[SYCL] Change the way of accessing single element in sycl::vec
Returning scalars from sycl::vec class as a swizzled vec causes problems
with template resolution and ambiguity in overload resolution. This
commit implements workaround for this problem. Scalars are returned by
value for const vec and by reference for vec. Assumed that it is
possible to get a reference the element of the vector, test is added to
check that behaviour is not changed.

Change-Id: I5b779152f668a29f3ef2d29e242b6ac42591d93b
Signed-off-by: Artur Gainullin artur.gainullin@intel.com
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/216259
Tested-by: xmain_checkin_sanity <lab_icls@iclgerrit.sc.intel.com>
Reviewed-by: Lin Jr, Chang-Sun <chang-sun.lin.jr@intel.com>