[SYCL] Enable useful (not random) output from stream
Pool of flush buffers is allocated in local memory. This pool contains
space for each work item in the work group. Each work item writes to
its own space (flush buffer), as a result output from different work
items is not mixed. Data is flushed to global buffer on endl, flush or
when kernel execution is finished.  Global buffer contains all output
from the kernel. Offset of the WI's flush buffer in the pool is
calculated only once in __init method. Call to this method is generated
by frontend.

In the current implementation user should explicitly flush data on the
host device. Data is not flushed automatically after kernel execution
because of the missing feature in the scheduler.

Change-Id: I883f38c0fb32cbc366f86cce61df21e10fc6fe9a
Signed-off-by: Artur Gainullin <artur.gainullin@intel.com>
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/216308
Tested-by: xmain_checkin_sanity <lab_icls@iclgerrit.sc.intel.com>
Reviewed-by: Lin Jr, Chang-Sun <chang-sun.lin.jr@intel.com>