Fix assert fail of ChannelPipeTransformation when using CoerceTypes
This issue follows the patch related to [CORC-5785].
Link: https://git-amr-2.devtools.intel.com/gerrit/#/c/214357/

The previous patch ignores the case when the members of a struct
contains pointers. Passing this type of struct to channel write
functions will cause assertion failure, while this should be legal
when writing such a struct to channel on FPGA. Therefore the assert
statements in the previous patch should be removed.

Two lit tests are added to test ChannelPipeTransformation pass
when coerced struct (with and without pointer member) is passed to
channel write functions.

Change-Id: I8c31be5e7f0bfb7b4be950c3d30a1fed94aae41a
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/216666
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Wenju He <wenju.he@intel.com>
Reviewed-by: Tianyang Zhu <tianyang.zhu@intel.com>
Reviewed-by: Dai, Chunyang <chunyang.dai@intel.com>