[VPlan][DA] Propagate stride information for select VPInstructions.
Currently DA sets stride for selects with external condition only if
the generated VectorShape is uniform. This patch extends DA to set
appropriate stride information even for cases when such selects
generate a strided VectorShape. The stride value is propagated only
if the operands being blended have the same stride.

Change-Id: I76e90b6221151d4b526ad73bdb68b74ad9b344a6
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/216804
Reviewed-by: Pushkar Ratnalikar <pushkar.v.ratnalikar@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Narayanaswamy, Ravi <ravi.narayanaswamy@intel.com>