[VPlan] Get rid of the 'Ptr' (Unit-stride pointer) Vector-shape descriptor.
'Ptr' shape is very specific and meant to identify unit-strided
pointers. This can be replaced, for all practical purposes, with a 'Str'
shape and a 'StrideVal'. The unit-stridedness can be determined with the
aid of an accompanying utility.

This change also adds some additional tests for,
*  Pointer-induction- File:linear_ptr_da.ll

In vplan_da_iv_shape.ll, I replace 'Strided' shape and an indeterminate
strideVal with 'Random' shape.

Change-Id: I855da6f329f10ecb3d1d1442a2b7255b641fb5fa
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/217630
Tested-by: xmain_checkin_sanity <lab_icls@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Kozhukhov, Sergey S <sergey.s.kozhukhov@intel.com>
Reviewed-by: Narayanaswamy, Ravi <ravi.narayanaswamy@intel.com>