CORC-6456 [SYCL][Driver] -fsycl -fintelfpga on Windows not triggering AOT
When using -fsycl -fintelfpga, which should imply the target triple for FPGA
AOT compilation was not setting the proper triple for Windows targets.  This
mismatch of targets was causing the device compilation to not be kicked off.
Updating the triple to the expected Windows value fixes this flow.

This is a cherry-pick of https://github.com/intel/llvm/pull/771

Change-Id: Ic69178e9cb7ec67470db0eab1bb632a76b3e823a
Signed-off-by: Michael D Toguchi <michael.d.toguchi@intel.com>
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/217606
Reviewed-by: Thatishetti, Meghana <meghana.thatishetti@intel.com>
Reviewed-by: Artem Gindinson <artem.gindinson@intel.com>
Tested-by: Artem Gindinson <artem.gindinson@intel.com>
Reviewed-by: Kurkov, Vasiliy A <vasiliy.a.kurkov@intel.com>