CORC-6457 [SYCL][Driver] Update -fintelfpga -fsycl-link behavior for unbundling objects
When unbundling objects when using -fsycl-link -fintelfpga, the expectation is
for the host and device parts of the object to be re-inserted into the FPGA
final device archive.  Updates to unbundling behavior causes only the host
objects to be extracted during the ubundling phase (this used to just return
the full fat object).  Now make sure the full object is passed to the archive
step.

This is a cherry-pick of https://github.com/otcshare/llvm/pull/628/commits/c5b3da7397ecac2f647cedd6f1a7f9fc4ba53d13

Change-Id: I2356ee08337efafc279e09516ae8a5b806022729
Signed-off-by: Michael D Toguchi <michael.d.toguchi@intel.com>
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/217639