Load OpenCL Plugin Library from Correct Path
`pluginPath` was NOT identical to plugin targets output directory
in in-tree build on Windows.

Usually, VS generator will append pre-configuration directory(Debug
or Release) to LIBRARY_OUTPUT_DIRECTORY, but it will NOT work with
in-tree build which use "Unix Makefiles" generator. Using `$<0:>`
to avoid the configuration-specific suffix and loading plugin from
current directory directly in any kind of builds.

This patch is related to CORC-6069.

Change-Id: I64a1cc844b1c8bdc1ec144746610fa2fb00eb130
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/217919
Reviewed-by: Ge Ren <ge.ren@intel.com>
Tested-by: Ge Ren <ge.ren@intel.com>