[VPlan][Merge loop exits] Updates the phi node of the exit block.
The cascaded if blocks algorithm connects each exit block to the right cascaded
if block. It also updates the phi nodes of the exit blocks so as to have the
right incoming blocks. In the case of while loops, we did not update the phi
node of the exit block that was processed last by the algorithm.

Change-Id: Iee58239be4c7081ab535ffb12e34478d8934b5c5
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/214834
Reviewed-by: Mitropoulou, Konstantina <konstantina.mitropoulou@intel.com>
Tested-by: Mitropoulou, Konstantina <konstantina.mitropoulou@intel.com>
Reviewed-by: Narayanaswamy, Ravi <ravi.narayanaswamy@intel.com>