[VecClone] Fix compfail associated with vector function returning a value from a load from a temp vs. alloca.
This patch addresses a hole in VecClone that is related to how return instructions
using the result of loads are handled. For return instructions that use the result
of a load from an alloca, VecClone reuses the alloca by widening it and returning it
as the result vector. Both the load and return are contained in the return block. The
existing store to this alloca remains in the loop so that each lane is able to store
the associated result. However, if there is a return from the result of a load from a
normal SSA temp, the load needs to stay in the loop and be followed by a new store
using an alloca generated for it and used as the return vector of the function. The
latter case was missing support in VecClone.

Examples - assume all instructions shown are in the entry block of a vector function:

1) .entry:
     %b.addr = alloca i32, align 4
     <snip, snip, snip> ...
     store i32 %inc, i32* %b.addr, align 4
     %1 = load i32, i32* %b.addr, align 4
     ret i32 %1

%b.addr is widened by VF and a new 'return' block is split containing both the load
and ret instructions. The store is placed within the newly formed loop body so that
each lane's value is correctly stored. This case is already supported.

2) .entry
     <snip, snip, snip> ...
     %arrayidx = getelementptr inbounds [4096 x i32], [4096 x i32]* @a, i64 0, i64 %idxprom
     %0 = load i32, i32* %arrayidx, align 4
     ret i32 %0

The load here does not come from an alloca, so VecClone must introduce a new widened alloca
for the return and insert a gep/store to this alloca within the loop. It must also make sure
to only put the 'ret' instruction in the 'return' block. This case was not handled and is
fixed with this patch. Basically, all the plumbing was already there, it was just that
VecClone had a hole by not distinguishing between the loads.

Change-Id: Ica930d5301bcf7183f5000d71ad1edca21a01922
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/218433
Reviewed-by: Masten, Matt <matt.masten@intel.com>
Tested-by: Masten, Matt <matt.masten@intel.com>
Reviewed-by: Narayanaswamy, Ravi <ravi.narayanaswamy@intel.com>