[SYCL][Driver] Use original archive when linking with an FPGA AOCX binary
When performing a final link step with -fintelfpga and given an AOCX device
archive, use the archive to link with as opposed to the unbundled host
objects.  The recent changes to the unbundler does not return the full
fat object, causing the host objects to not be 'found' during the final host
unbundling.

This is a cherry-pick of https://github.com/otcshare/llvm/pull/636.

Change-Id: I5dc3777b70b761bee80069a2be50b0e52b493d7d
Signed-off-by: Michael D Toguchi <michael.d.toguchi@intel.com>
Signed-off-by: Artem Gindinson <artem.gindinson@intel.com>
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/218521
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Kurkov, Vasiliy A <vasiliy.a.kurkov@intel.com>