[DTrans] [CMPLRLLVM-10825] Improve DTrans local pointer analysis traces by intermixing type resolution into IR dump
This adds a debug trace to help analyze the type information collected
about pointers by the DTrans local pointer analyzer by printing the
complete results of the LPA intermixed with an IR dump. The current
dumps print information as instructions are analyzed, but this is
not easily mapped to the IR of a funciton.

This change does not have an effect on the compiler generated code.

Change-Id: I0836744559329c9f9c63363c3277d5006ea2e390
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/218868
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Reviewed-by: Narayanaswamy, Ravi <ravi.narayanaswamy@intel.com>