[X86] Enable use of MSVC style __cpuid and __cpuidex outside of MSVC compatibility to match icc.
This is a copy of the function signatures from intrin.h with the body
taken from cpuid.h due to the rbx issue that may exist on Linux
that we work around in the implementation.

I've qualified this with !defined(_MSC_VER) so the version
in intrin.h is still used when being compatible with MSVC.
I've also checked for __cpuid already being defined to avoid
a conflict with the user also importing cpuid.h.

I've also expanded the INTEL_CUSTOMIZATION region in
immintrin.h to cover the FEATURE bits that are customized
but weren't marked.

Fixes CMPLRLLVM-10818

Change-Id: I609d83984ce3f958beee0cf3066f178884b00048
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/218905
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Luo, Yuanke <yuanke.luo@intel.com>
Reviewed-by: Narayanaswamy, Ravi <ravi.narayanaswamy@intel.com>