Timeout AffinitizeThreads and pin thread upon registration
If user's master thread doesn't join in OpenCL RT's tbb arena,
AffinitizeTask hangs because barrier never becomes 0 and all threads in
the arena are never released (CORC-6046). This is fixed by setting
thread affinity in on_scheduler_entry when tbb thread participates to do
a job for the first time.
AffinitizeThreads can be removed after the change of setting affinity in
on_scheduler_entry. However, setting affinity could happen as late as
the first kernel execution and thus introduces overhead to kernel time.
Therefore, we set timeout in AffinitizeThreads task. If it times out,
the threads that don't join this task will have their affinities set
when participating to do a job later.

Change-Id: I062bcc3d0d5bc185c61570504eea1dec353e5298
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/217462
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Romanov, Vlad <vlad.romanov@intel.com>
Reviewed-by: Saito, Hideki <hideki.saito@intel.com>
Reviewed-by: Dai, Chunyang <chunyang.dai@intel.com>