[VPlan] - Change to identify  scalar-private load/store/gep as unit-stride.
We still end up generating a GEP which return the vector of pointers to
all the elements of the widened-scalar. In case of a load which directly
uses alloca, we use the wide base-ptr directly. In case of a load via a
GEP, we see an extract-element on the GEP vector of pointers and an
additional bitcast. This instruction-sequence can be improved in the future.

Change-Id: I7fbdbc58af8989b91375cda6384de9833918059d
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/218906
Tested-by: xmain_checkin_sanity <lab_icls@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Karthik Senthil <karthik.senthil@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Masten, Matt <matt.masten@intel.com>
Reviewed-by: Narayanaswamy, Ravi <ravi.narayanaswamy@intel.com>