[VPlan][SVA] Introduce analysis to compute Scalar/Vector nature of VPInstructions.
This patch adds a new analysis in VPlan to populate the data
structure that tracks Scalar/Vector nature of a VPInstruction.
The computation and propagation of an instruction's nature is
implemented via an algorithm that works on def->use chains and
considers divergence information from DA. The analysis has the
flexibility to be run for multiple iterations, with the only
prerequisite of having correct DA results to query.

JIRA : CMPLRLLVM-9316

Change-Id: I5e3bb6d3f60a6674300caea6181856cf5c962f3f
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/219308
Reviewed-by: Karthik Senthil <karthik.senthil@intel.com>
Tested-by: Karthik Senthil <karthik.senthil@intel.com>