[VPlan] Remove VPOCodeGen::emitMinimumIterationCountCheck, we have emitVectorLoopEnteredCheck
We are generating the same (semantically, albeit with slightly different
instructions sequence) check twice. No need to do that - remove one of them.

I decided to remove the one added first, because it would affect less tests -
all the instruction added to the preheader are effectively added to the
successor of the second check - let's keep that unchanged.

Change-Id: I8730864b604bd895f705f947ffcedf1f4080b94b
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/219316
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Guggilla, Satish <satish.guggilla@intel.com>
Reviewed-by: Narayanaswamy, Ravi <ravi.narayanaswamy@intel.com>