[X86] Add intrinsics for stac/clac/lgdt/sgdt instructions to match icc.
sgdt/lgdt seem to vary the number of bytes read/written
based on the operating mode of the CPU and the intrinsics
take a void pointer. So I've casted it to a packed struct
that has a size of 10 bytes in 64-bit mode and 6 bytes
otherwise. I've also marked it as clobbering memory so it
will be a memory barrier for other operations. These
instructions mess with the descriptor table that controls
segments so this seemed safest. Might only need this on
lgdt, but I put it on sgdt as well. These instructions
aren't accessible to normal applications and could only be
used in ring 0 so its probably good to be safe.

I've also made stac/clac memory barrier as well. Those
instructions control whether the kernel can access user
pages so we wouldn't want loads/stores being moved
across them.

Change-Id: Ifa2cfe92a1e3bcd8185f7ff68dff924b908fe1e3
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/219556
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Luo, Yuanke <yuanke.luo@intel.com>
Reviewed-by: Narayanaswamy, Ravi <ravi.narayanaswamy@intel.com>