[VPlan] This patch fixes CMPLRLLVM-10876 - compfail for spec2006all464C/464@opt_O2_outer_vec.
The compfail happened because mergeLoopExits() in HCFGBuilder assumed that successor blocks
of newly introduced cascaded if blocks (exit blocks of loops) were always contained in a
parent loop. In this case, one of these successor blocks was an exit block for the outermost
loop. This resulted in a nullptr dereference. This patch checks to make sure that these
blocks are contained within a loop, and if not, does not try to add the block to a parent
loop.

Change-Id: Ib15535178e7aa2310fc7c914fc3af1f772d3d3d6
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/219801
Reviewed-by: Kozhukhov, Sergey S <sergey.s.kozhukhov@intel.com>
Tested-by: xmain_checkin_sanity <lab_icls@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Narayanaswamy, Ravi <ravi.narayanaswamy@intel.com>