[VPlan] Fix for incorrect replacement of induction start and induction step
During generation of VPInstructions for VPEntities the VPInductionInit and
VPInductionInitStep are used to replace induction start and step in the loop body.
The current implementation does not work for cases when start and step are equal
constants. It replaces both of them with VPInductionInit. Really, we need to
replace them more accurately, targeting starting phi (for init) and the update
instruction (for step).

JIRA: CMPLRLLVM-10900.

Change-Id: I2e5f7ea634cf9a8210d5c45f238188ff900bc040
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/220092
Reviewed-by: Kozhukhov, Sergey S <sergey.s.kozhukhov@intel.com>
Tested-by: Kozhukhov, Sergey S <sergey.s.kozhukhov@intel.com>