Fixed an issue with unchanged command queue properties in API tracing
Root cause: functions clCreateCommandQueue and
clCreateCommandQueueWithProperties receive "properties" argument
that is passed to API tracer, but this argument is not passed to
the next level (to internal function CreateCommandQueue) since
another "internal" variable is passed to it. So any modifications
of properties don't take an effect.

Fix: just pass "internal" properties variable to API tracer.

Change-Id: I107c0d83fd2ece692c431ee2e85c1a562240c8df
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/220669
Reviewed-by: Dai, Chunyang <chunyang.dai@intel.com>
Tested-by: Gorshkov, Anton V <anton.v.gorshkov@intel.com>