Allow function and module dumps in debug modes
This change removes a restriction that was preventing function and
module text dumps when we have built with a configuration that allows
debug output. These were originally added as trip wires so that we would
notice if the pass pipeline was reordered in a way that could lead to IR
printing after proprietary optimizations had been performed. The
BitcodeWriter still checks for that path, and the AsmWriter code is
excluded from product release builds.

This fixes CMPLRLLVM-10971.

Change-Id: I7e20c20e68ab06ef0952e719e7b4034050ac3dec
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/221087
Tested-by: xmain_checkin_sanity <lab_icls@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Reviewed-by: Narayanaswamy, Ravi <ravi.narayanaswamy@intel.com>