[DTrans] [CMPLRLLVM-10883] Improve analysis to handle memset calls that start on padding regions
This change is to improve the DTrans analysis to enable it being run at the
time of the HIR loop passes to identify functions that access specific structure
fields.

Currently, the IR at this phase sets a structure with the "Bad pointer manipulation"
safety bit if a GEP instruction is used to get the address within a structure which
corresponds to padding between fields. This change relaxes that safety check for the
case where the address is used for a call to memset, and extends the analysis
of memset calls to allow for starting the memset on the padding address.

Change-Id: I1aa8d77d6dd34015d1c52a6dcefc00c4a6b22461
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/221449
Reviewed-by: Narayanaswamy, Ravi <ravi.narayanaswamy@intel.com>
Tested-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>