[VPlan] Temporarily don't allow vectorization of masked fabs intrinsics.
A hot loop in 544.nab was found to contain masked fabs intrinsic
call, which after vectorization resulted in performance regressions
on AVX2 architecture. This patch temporarily disables vectorization
of loops containing these intrinsics as a workaround until VPlan
Cost Model and AllZero bypass are uplifted to correctly model this
hot loop. The cl::opt to disable this workaround is
-vplan-vectorize-masked-fabs.

Change-Id: I167cd0c4f8930d715d25e1bd95853eb7b838c15d
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/221538
Reviewed-by: Kozhukhov, Sergey S <sergey.s.kozhukhov@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Narayanaswamy, Ravi <ravi.narayanaswamy@intel.com>