[VPlan][LoopCFU][1/2] Refactor to work without non-loop regions simplification
This commit doesn't change the output of the transformation for now, but is
required to be implemented for a subsequent disabling of extra blocks splitting
that was initially introduced to form more SESE regions.

As such, this change has to be tested, but it will be covered by the existing
LIT tests once that second commit lands (we'd have crashes without this
modification otherwise).

Change-Id: I4de9176bb2a7a17cdc8a8872c38041ea4bee00f0
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/223666
Reviewed-by: Elovikov, Andrei <andrei.elovikov@intel.com>
Tested-by: Elovikov, Andrei <andrei.elovikov@intel.com>