[VPlan][DA] Fix vector shapes of divergent users when instruction is updated.
Currently VPlan's DA does not re-process already divergent instructions
that are in the worklist. This could lead to incorrect Vector Shapes
computation, for example -

[Unit Stride] %div.inst = <..>
[Uniform]     %upd.inst = <..>
[Unit Stride] %div.user = add %div.inst, %upd.inst

In the next iteration of worklist, if %upd.inst is updated to Random
then %div.user should also be updated to reflect the new shape. This
patch fixes DA to ensure that all users of updated instructions are
revisited.

Change-Id: I6be39806c034545c2f60559c74e741e79cab23da
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/222170
Reviewed-by: Guggilla, Satish <satish.guggilla@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Masten, Matt <matt.masten@intel.com>
Reviewed-by: Narayanaswamy, Ravi <ravi.narayanaswamy@intel.com>