[SLP][NFC] Replace Intel customized shallow score calculation with the community version.
The change is further move toward replacing Intel customized Look-Ahead score
calculation stuff with community routines. Scores are adjusted a bit so that
the change would not affect current Multi-Node grouping behavior.
That makes the change not exactly a no-op change but it is still intended
to remain functionally equivalent.

Change-Id: I090d050d6332eea752439b0de44016b36578a07a
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/222332
Reviewed-by: Dmitriev, Valery N <valery.n.dmitriev@intel.com>
Tested-by: Dmitriev, Valery N <valery.n.dmitriev@intel.com>
Reviewed-by: Narayanaswamy, Ravi <ravi.narayanaswamy@intel.com>