Implement cl_intel_global_variable_pointers ext
New api clGetDeviceGlobalVariablePointerINTEL returns pointer to a
global variable (aka program scope variable) of OpenCL program.

The returned pointer is represented as a USM device pointer. USM
wrappers for all the global variables are created after device build
of program. USM wrappers are released in clReleaseProgram because
lifetime of both global variables and their USM wrappers are tied to
the program.

Metadata opencl.global_variable_total_size is removed since it can now
be calculated from the map of global variables and sizes.

Change-Id: I11cf9dc8d483c55a2a324331cb9e0b290b3ca4e5
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/219144
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Dai, Chunyang <chunyang.dai@intel.com>