Implementation of clSetProgramSpecializationConstant API
The API is available via clGetExtensionFunctionAddressForPlatform().

In order to report errors required by the OpenCL spec, we need to
get information about ids of specialization constants in the module
as well as their size. To accomplish that, the implementation adds
GetSpecConstInfo method to the frontend API. GetSpecConstInfo is
called only first time we call clSetProgramSpecializationConstant
for the given program. The info returned by GetSpecConstInfo is
cached in ProgramWithIL object and is used for validation of
arguments passed to clSetProgramSpecializationConstant API.

Specialization constants provided via
clSetProgramSpecializationConstant are also accumulated in
ProgramWithIL object and eventually they all together are passed to
ClangFECompilerParseSPIRVTask::ParseSPIRV() when the user calls
clBuildProgram.

Change-Id: I5cadadddfca3f00e8046eeb13dc5710eaee20dfc
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/231249
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Sotkin, Alexey <alexey.sotkin@intel.com>