Implementation of clSetProgramSpecializationConstant API
The API is available via clGetExtensionFunctionAddressForPlatform().

In order to report errors required by the OpenCL spec, we need to
get information about ids of specialization constants in the module
as well as their size. To accomplish that, the implementation adds
GetSpecConstInfo method to the frontend API. GetSpecConstInfo is
called only first time we call clSetProgramSpecializationConstant
for the given program. The info returned by GetSpecConstInfo is
cached in ProgramWithIL object and is used for validation of
arguments passed to clSetProgramSpecializationConstant API.

Specialization constants provided via
clSetProgramSpecializationConstant are also accumulated in
ProgramWithIL object and eventually they all together are passed to
ClangFECompilerParseSPIRVTask::ParseSPIRV() when the user calls
clBuildProgram.

Change-Id: If45c62817152e612eacaade9005f41362e45dce7
Signed-off-by: Alexey Sotkin <alexey.sotkin@intel.com>
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/227945
Reviewed-by: Romanov, Vlad <vlad.romanov@intel.com>