[NFC][VPlan] First step of VPRecipe removal from VPlan sources - make VPInstruction derive from VPUser and iplist.
The changes here are the first step in removing VPRecipes from VPlan sources.
As the first step, VPInstruction has been modified to derive from VPUser and
iplist. Earlier VPInstruction was derived from VPUser and VPRecipeBase.

- this changeset does not get rid of all redundant uses of isa<VPInstruction>,
cast<VPInstruction>. I have removed the obvious uses that were touched as part
of these changes. I will remove other redundant uses as a separate changeset.
- VPRecipes removal will follow as a separate changeset.

This changeset needs to go in after removal of the old predicator sources that
is currently in flight.

Change-Id: I0cdbfb24f4b0a18dbdb38b6a88a98779689a0724
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/223036
Reviewed-by: Guggilla, Satish <satish.guggilla@intel.com>
Tested-by: Guggilla, Satish <satish.guggilla@intel.com>