Fix compile error for handler::copy with -fsycl-unnamed-lambda
Some `copy` methods in handler are implemented via `parallel_for` for
host device and hence, we need to have `KernelInfo` struct (provided by
integration header) to compile them.
This effectively creates cyclic dependency and we already have some kind
of a workaround for it for regular mode (without -fsycl-unnamed-lambda).

This patch applies the same solution for unnamed lambda mechanism.

Signed-off-by: Alexey Sachkov <alexey.sachkov@intel.com>
Change-Id: Idc7ce30486f189fafc24c46c0559dac4332bce61
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/223123
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Chen, Helen C <helen.c.chen@intel.com>
Reviewed-by: Jayashankar, Prema <prema.jayashankar@intel.com>