Fix compile error for handler::copy with -fsycl-unnamed-lambda
Some `copy` methods in handler are implemented via `parallel_for` for
host device and hence, we need to have `KernelInfo` struct (provided by
integration header) to compile them.
This effectively creates cyclic dependency and we already have some kind
of a workaround for it for regular mode (without -fsycl-unnamed-lambda).

This patch applies the same solution for unnamed lambda mechanism.

Signed-off-by: Alexey Sachkov <alexey.sachkov@intel.com>
Change-Id: I7cd447e015d31547d2602ff420bebdadd095a3ad
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/223121
Reviewed-by: Romanov, Vlad <vlad.romanov@intel.com>
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Huang, Zao <zao.huang@intel.com>