CMPLRLLVM-11101 - Use calls instead of invokes in OpenMP regions
The OpenMP spec says something like this for directives:

A throw executed inside a region must cause execution to resume within
the same region, and the same thread that threw the exception must
catch it.

This change forces a call instead of an invoke for calls in an OpenMP
region when the call is not inside a try-block.

Change-Id: Ibc9e08879f7730edb4bc09906ab51cb6e55d0e50
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/223241
Reviewed-by: Yu, Jennifer <jennifer.yu@intel.com>
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Narayanaswamy, Ravi <ravi.narayanaswamy@intel.com>