[CMPLRLLVM-9909] Add an early row-wise multiversioning pass
This is the first change in a ~3 change series to fully implement this
pass: it implements the analysis to determine when and how to apply the
transformation, but does not implement any code transformation yet. That
will be added in later changes.

What this version implements:

 - Running the row-wise multiversioning pass via opt
 - Analysis for determining multiversioning candidates, including local
   legality checks and non-local DTrans queries to determine possible
   array values

What will be implemented in future changes:

 - Adding this pass to the pass schedule
 - The multiversioning transform itself
 - The extension to support non-invariance in outer loop levels via
   arrays

Change-Id: Ib5e367bd9882ccd27c3d7f31c8742e87fabec2c5
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/217788
Tested-by: xmain_checkin_sanity <lab_icls@iclgerrit.sc.intel.com>
Reviewed-by: Chawla, Pankaj <pankaj.chawla@intel.com>
Reviewed-by: Narayanaswamy, Ravi <ravi.narayanaswamy@intel.com>