[LLD] CMPLRLLVM-11158: Support for combined LTO objects for GCC9+ [XMAIN]
We use the relocatable option (-r) to create a real object out of a
GNU LTO object. This rule changed for GCC/G++ 9+. If a relocatable object
is created from LTO objects, the IR will be preserved and a new GNU LTO
object will be created. A new flag (-flinker-output=nolto-rel) was added
to generate a real ELF object out of LTO objects. This change set basically
tries to identify if the flag is supported in GCC. If so the add it into
the command used by GCC.

A test case will be submitted to the TC system to catch the multiple
versions of GCC.

Change-Id: I65f212774276102451779fd58a49e4e9f78c77f3
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/223515
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Narayanaswamy, Ravi <ravi.narayanaswamy@intel.com>