[CMPLRLLVM-10533] Add support for loop-based bound ranges in SCEV.
This modifies scalar evolution to support a maximum bound on a value if it is
incremented or not incremented every loop iteration, such as in this loop:

int x = 0;
for (int i = 0; i < 60; i++) {
  if (unpredictable())
    x++;
}
// x is now known to be in the range [0, 60).

Unfortunately, enabling this change for all SCEV range queries causes failures
in LoopOpt. For the moment, this method is only provided to consumers who want
more accurate range information, such as the upcoming array use analysis.

Change-Id: I7d0bf8963da3593f68ee0898a24f5231b58e0f11
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/224558
Reviewed-by: Chawla, Pankaj <pankaj.chawla@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Topper, Craig <craig.topper@intel.com>