[SYCL][FPGA] Temporarily restore old version of IVDep loop attribute
When the switch to the array-aware version of the attribute was
taking place, we didn't consider additional work/time required
for the FPGA BE to accept the new LLVM IR layout. Meanwhile, having
some support of the ivdep attribute is still obligatory.

This patch provides backwards compatibility for ivdep after
[6a283ae8d5254]. A mere revert wouldn't be good enough, as the latter
patch has also introduced the infrastructure for handling dependent
arguments in such attributes, which has been re-used since. Besides,
retaining the new version of ivdep allows for seamless FPGA BE
implementation and testing.

Consequently:
1) the new, array-aware version of the attribute was temporarily
renamed into [[intelfpga::ivdep_exp]]
2) the legacy ivdep implementation was restored under the original
name [[intelfpga::ivdep]]
3) the dependent argument handling was added for the legacy variant
of the attribute using the aforementioned infrastructure, as this
support is expected regardless of the LLVM IR layout

Once the FPGA BE and the SPIR-V Translator fully support the new variant
of ivdep, this patch must be reverted. This will finalize the LLVM IR
layout switch.

Change-Id: I28af649e0c3a925031c7e92d66d5293905a3df0d
Signed-off-by: Artem Gindinson <artem.gindinson@intel.com>
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/224885
Reviewed-by: Sidorov, Dmitry <dmitry.sidorov@intel.com>
Reviewed-by: Topper, Craig <craig.topper@intel.com>