[InstCombine] Descale sums and differences
E.g., when asked to descale (X + Y), return Descale(X) + Descale(Y).
This succeeds only when both X and Y can be descaled. Binary Or is
treated as addition when appropriate. Subtraction is also supported.

In order to avoid making unnecessary changes to the IR and potentially
creating infinite loops via InstCombine, this also changes Descale to
accept an option making it a non-destructive test. This feature is used
by Descale itself to determine if both X and Y can be descaled.

Change-Id: I5a9e53807bc72d0a8558c804fac4160af2b1ef02
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/226314
Reviewed-by: Creech, Timothy M <timothy.m.creech@intel.com>
Tested-by: Creech, Timothy M <timothy.m.creech@intel.com>