[CMPLRLLVM-11023] [FPGA] Implementation of the pragma loop_fuse scoped attribute
The specification of the user interface and the compiler interface are provided
in the JIRA:
https://jira.devtools.intel.com/browse/CMPLRLLVM-11023

This patch implements the scope pragma as a new type of attribute: LoopFuseAttr.
It re-uses the infrastructure in ParsePragma.cpp to recognize and parse the
pragma and creates an AttributedStmt of the following structure for the scope
(CompoundStmt) to which the pragma applies:
  |-AttributedStmt
  | |-LoopFuseAttr
  | | `-CompoundStmt
  |     |
  |     |

The attribute has two arguments that correspond to the two possible clauses
that can be applied to the pragma:
- Optional Integer parameter (default to 1) for the `depth` clause.
- A String parameter for the 'independent' clause.

In Semantic Analysis, LoopHint infrastructure is re-used to verify the validity
of the integer-type argument to the `depth` clause.

For CodeGen, a similar kind of approahc is used to generate the directive.entry
and directive.exit as with the ivdep pragma.

Regression tests for Parsing/Sema and CodeGen are added in:
clang/test/CodeGenIntelHLS/loop_fuse.cpp
clang/test/CodeGenOpenCL/intel-fpga-loop_fuse.cpp
clang/test/SemaIntelHLS/loop_fuse.cpp

Change-Id: I99dc521cf7fa9e3fb6665e24011cea256edec835
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/229395
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Andrews, Elizabeth <elizabeth.andrews@intel.com>
Reviewed-by: Narayanaswamy, Ravi <ravi.narayanaswamy@intel.com>