Avoid deleting cleanuppads with liftime.end
This works around a problem where we were creating non-SSA IR after
deleting a cleanup pad with a PHI node that did not dominate the block
we were sinking it into. This was caused by a PHI node in the cleanup
pad whose only user was a lifetime.end intrinsic in the cleanup pad.

See CMPLRLLVM-11251 for more discussion of the problem.

Change-Id: Ibd6f8eaa8457359d7af797864f04d4b9d782b69d
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/225191
Tested-by: Kaylor, Andrew <andrew.kaylor@intel.com>
Reviewed-by: Topper, Craig <craig.topper@intel.com>