[VPlan] Implement pumping for SVML calls in LLVM-IR vector codegen.
This patch introduces an optimization for vectorizing SVML calls,
when the call cannot be widened for a chosen VF, by pumping them
with a lower VF. This is achieved by breaking down the call
arguments to the selected lower VF context and using them to emit
multiple vector calls determined by pumping factor. These call
results are subsequently combined back to be used in the original
VF context.

JIRA: CMPLRLLVM-10655

Change-Id: I389e78f798bb8c16a598c9d2666cc16c829bf6b7
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/225418
Reviewed-by: Karthik Senthil <karthik.senthil@intel.com>
Tested-by: Karthik Senthil <karthik.senthil@intel.com>
Reviewed-by: Narayanaswamy, Ravi <ravi.narayanaswamy@intel.com>