[FPGA] Serialize execution of instances of the same kernel in OOO queue
If we have the same kernel being enqueued on FPGA emulator several
times - we don't want two or more instances of this kernel being
executed concurrently. That is needed in some cases, if the kernel
contains SYCL/OpenCL pipes/channels to prevent possible race condition
because of several instances of the same kernel are using the same
pipe/channel from different threads (that leads to program hang).

None of the specifications guarantee this execution order, but FPGA
hardware works that way and so the emulator shall mimic the
hardware's behaviour.

To achieve this, we create a map of events to kernels. If execution
of a kernel is finished, we can enqueue it one more time. Otherwise,
we wait until the kernel is finished.

Please note, that for SYCL we currently have 1:1 mapping of Q.submit
on enqueue_kernel OpenCL API function calls with the order of kernels
to be enqueued saved from SYCL to OpenCL. But it isn't necessarily
will be true in the future. When/if this behaviour is changed - this
patch must be reverted and actually I don't know if kernel
serialization is possible in this case.

Following restrictions are applied for this approach:
  - Don't serialize autorun kernels - they are running from a start of a
    program till its termination;
  - Multi-device programs are not handled by this patch.

Change-Id: I98c07d0cbf4a5451f805494590bf4ff4ec514a6f
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/224242
Tested-by: Sidorov, Dmitry <dmitry.sidorov@intel.com>
Reviewed-by: Mikhail Lychkov <mikhail.lychkov@intel.com>
Reviewed-by: Dai, Chunyang <chunyang.dai@intel.com>