CMPLRLLVM-9946: cleaned up master thread check generation.
This is just a cleanup for master thread check generation.
The ultimate solution for specACCEL/503 is to get rid of
the master thread check completely, which will be done
in another change-set.
Anyway, this clean-up provides 1.3x speed-up for specACCEL/503
for some reason, so it should be good in general.

Change-Id: I8a753ed20a503772b282e2abebea1dd5d95d2fdc
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/225588
Reviewed-by: Su, Ernesto <ernesto.su@intel.com>
Reviewed-by: Tian, Xinmin <xinmin.tian@intel.com>
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Topper, Craig <craig.topper@intel.com>