[intel_subgroup] Fix block read/write segmentation fault
Previously, we use data = *(vectype*)pSrc for block read and use
*(vectype*)pDest = data for block write. This will indicate the
compiler to generate aligned load/store IR.
In OpenCL C SPEC:
6.1.5. Alignment of Types. It says "A data item declared to be a
data type in memory is always aligned to the size of the data type
in bytes.";
6.2.5. Pointer Casting. It says "Pointers to old and new
types may be cast back and forth to each other. Casting a pointer
to a new type represents an unchecked assertion that the address is
correctly aligned"
So when the address is actually not aligned, segmentation fault
will occur. We can use vload/vstore to do the block read/write,
it produces load/store IR aligned to 1.

Change-Id: I99f6eb409f4b09bb5998b2cc04cd4ad4b59018ca
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/226011
Tested-by: wuwei <wei.wu@intel.com>
Reviewed-by: Wenju He <wenju.he@intel.com>
Reviewed-by: Tyurin, Evgeniy <evgeniy.tyurin@intel.com>
Reviewed-by: Dai, Chunyang <chunyang.dai@intel.com>