[FPGA] Add pass to remove call for "__cxa_atexit"
This patch is a workaroud for support linking c++ code
with opencl kernel code. The c++ code has global ctors/dtors
which may involve dynamic library and leads to link error
for "__dso_handle". As a temporialy workaroud, we directly
remove invoke of __cax_atexit.

We will provide a official solution later.

Change-Id: I94415f7cec57c7ee4f688fd7fe8826445d934d6e
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/226145
Reviewed-by: Dai, Chunyang <chunyang.dai@intel.com>
Tested-by: Dai, Chunyang <chunyang.dai@intel.com>