[NFC][VPlan][1/4] Introduce VPCallInstruction to represent calls in VPlan.
This patch adds a new specialization for the VPInstruction class to
represent Call instructions in VPlan CFG. The motivation comes from
the need to explicitly represent call vectorization decisions which
can be used by various components in VPlan to understand the nature/
behavior of a given Call instruction. The patch also introduces
CallVecProperties and CallVecScenarios data structure used to record
the above mentioned decisions. This is the first patch in the series
for the task to centralize call vectorization decisions in VPlan.

Change-Id: I092d886a43b600f252fc5fca78681baf24c06fe8
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/228675
Reviewed-by: Dmitriev, Valery N <valery.n.dmitriev@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Kozhukhov, Sergey S <sergey.s.kozhukhov@intel.com>
Reviewed-by: Dai, Chunyang <chunyang.dai@intel.com>